requirements.txt
aiogram==3.12.0
pytesseract
pdf2image
pillow
PyPDF2
sympy
python-magic-bin==0.4.14
